refactor(apple): consolidate root views into one, simplify where possible
